add run, init, pause and message sags tests
